add const qualifier to pink_regset
There are many functions with prefix pink_vm_*, pink_read_* and
pink_write_* which accept a pointer to a struct pink_regset as
argument but do not modify it. Add const qualifier to the regset
argument of these functions.

In addition add the nonnull compiler attribute to pink_regset
arguments as it is never valid to have a NULL pointer for this
argument.

Signed-off-by: Ali Polatel <alip@exherbo.org>